WIP: filesze of history entry seems not be corret

+++ src/HAP/plugins/BME280/HAPPluginBME280.hpp
@@ -93,7 +93,7 @@ public:
 		return _humidityAverage.getAverage();
 	}
 
-	inline float getAveragedPressureValue(){
+	inline uint16_t getAveragedPressureValue(){
 		return _pressureAverage.getAverage();
 	}
